+++ chainforge/react-server/src/SplitNode.js
@@ -12,51 +12,11 @@ import { processCSV, deepcopy, deepcopy_and_modify, dict_excluding_key, toStanda
 
 import { fromMarkdown } from "mdast-util-from-markdown";
 import StorageCache from './backend/cache';
-
-const formattingOptions = [
-  {value: "list",    label:"- list items"},
-  {value: "\n",   label:"newline \\n"},
-  {value: "\n\n", label:"double newline \\n\\n"},
-  {value: ",",    label:"commas (,)"},
-  {value: "code",  label:"code blocks"},
-  {value: "paragraph",  label:"paragraphs (md)"},
-];
-
-const truncStr = (s, maxLen) => {
-  if (s.length > maxLen) // Cut the name short if it's long
-      return s.substring(0, maxLen) + '...'
-  else
-      return s;
-};
-const tagMetadataWithLLM = (input_data) => {
-  let new_data = {};
-  Object.entries(input_data).forEach(([varname, resp_objs]) => {
-    new_data[varname] = resp_objs.map(r => {
-      if (!r || typeof r === 'string' || !r?.llm?.key) return r;
-      let r_copy = JSON.parse(JSON.stringify(r));
-      r_copy.metavars["__LLM_key"] = r.llm.key;
-      return r_copy;
-    });
-  });
-  return new_data;
-};
-const extractLLMLookup = (input_data) => {
-  let llm_lookup = {};
-  Object.entries(input_data).forEach(([varname, resp_objs]) => {
-    resp_objs.forEach(r => {
-      if (typeof r === 'string' || !r?.llm?.key || r.llm.key in llm_lookup) return;
-      llm_lookup[r.llm.key] = r.llm;
-    });
-  });
-  return llm_lookup;
-};
-const removeLLMTagFromMetadata = (metavars) => {
-  if (!('__LLM_key' in metavars))
-    return metavars; 
-  let mcopy = JSON.parse(JSON.stringify(metavars));
-  delete mcopy['__LLM_key'];
-  return mcopy;
-};
+import { formattingOptions } from './backend/utils';
+import { tagMetadataWithLLM } from './backend/utils';
+import { extractLLMLookup } from './backend/utils';
+import { removeLLMTagFromMetadata } from './backend/utils';
+import { truncStr } from './backend/utils';
 
 /** Flattens markdown AST as dict to text (string) */
 function compileTextFromMdAST(md) {
